Add files via upload
This is a Movie Recommendation System based on Content Based Filtering Method. This project uses a dataset moviesdata.csv having around 4000 entries. The data is first cleaned and then with the help of cosine similarity, we find the similarity score of the movies with respect to the users favourite movie. Then we recommend  20 similar movies from the dataset.